tor: remove circular import
